--- a/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.empty_commit.approved.json
+++ b/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.empty_commit.approved.json
@@ -26,6 +26,8 @@
     "HasMultiLineMessage": false,
     "HasNotes": false,
     "IsArtificial": false,
+    "IsStash": false,
+    "ReflogSelector": null,
     "HasParent": true,
     "FirstParentId": {
         "IsArtificial": false