--- a/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.no_subject.approved.json
+++ b/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.no_subject.approved.json
@@ -32,6 +32,8 @@
     "HasMultiLineMessage": false,
     "HasNotes": false,
     "IsArtificial": false,
+    "IsStash": false,
+    "ReflogSelector": null,
     "HasParent": true,
     "FirstParentId": {
         "IsArtificial": false