--- a/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.normal.approved.json
+++ b/UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.normal.approved.json
@@ -32,6 +32,8 @@
     "HasMultiLineMessage": true,
     "HasNotes": false,
     "IsArtificial": false,
+    "IsStash": false,
+    "ReflogSelector": null,
     "HasParent": true,
     "FirstParentId": {
         "IsArtificial": false