--- a/UnitTests/GitCommands.Tests/TestData/RevisionReader/reflogselector.bin
+++ b/UnitTests/GitCommands.Tests/TestData/RevisionReader/reflogselector.bin
@@ -1,0 +1,15 @@
+0c350e248500e4e893e6d17dcc61591cc1cd64e71bcb98ebe109b0a9a569b6c827d55f97730cdb247b89abaf240ec4e057a2304659cbde7b1e7bb68e 7b89abaf240ec4e057a2304659cbde7b1e7bb68e 7b89abaf240ec4e057a2304659cbde7b1e7bb68e
+1624479586
+1624479586
+Gerhard Olsson
+gerhardol@users.noreply.github.com
+Gerhard Olsson
+gerhardol@users.noreply.github.com
+stash{0}
+wip tests
+
+second line
+third line before multiple whitespace
+		
+
+