--- a/UnitTests/GitUI.Tests/UserControls/FilterInfoTests.cs
+++ b/UnitTests/GitUI.Tests/UserControls/FilterInfoTests.cs
@@ -20,7 +20,6 @@ namespace GitUITests.UserControls
         public void SetUp()
         {
             AppSettings.ShowGitNotes = false;
-            AppSettings.ShowLatestStash = true;
         }
 
         [Test]