… GitUI/UserControls/RevisionGrid/RevisionGridControl.cs
… IntegrationTests/UI.IntegrationTests/CommandsDialogs/FormBrowseTests.cs
… UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.reflogselector.approved.json
… UnitTests/GitCommands.Tests/RevisionReaderTests.TryParseRevision_test.ForScenario.subject_no_body.approved.json
… UnitTests/GitCommands.Tests/RevisionReaderTests.cs