Resolves #5371 (and #5785
Fixes #9873

## Proposed changes

Show stash commits in the revision grid, with menu options to pop and drop stashes.

A stash is implemented in Git with 3 internal commit.
Some more explanations in GitUI\UserControls\RevisionGrid\RevisionGridControl.cs:988
The implementations I have seen do not expose the third commit with untracked files.
The PR includes this "untracked" commit (empty "untracked" are filtered) for the first 10 stashes (see Performance below for details).
This setting is read from the settings but no setter is exposed, mostly because there is no natural place to show it (maybe in View is the best, but other similar settings are in General, that is considered full).

The implementation by default only show the stashes where the parents are shown in the grid.
So for instance if you stash and then rebase then the stash is not shown (but see below).
Unless reflogs are shown, only stashes that are related to commits visible in the grid is shown.
(Another difference with reflog is that stashes are presented in Git order, the "inserted" stashes are always shown related to the main commit).

The label for stash commits is partly for this reason set to the reflog selector like "stash@{0}" (required to pop/drop stashes) (another reason is to not get lost too easy when looking at commits, especially when many stashes are created for the same commit).
(For instance GitKraken shows these commits.)
FormStash can still be used to see these, they could also be added to the side panel (similar to SourceTree) (but the contents would not be shown). See #5460.
(Sublime merge shows both in grid and sidepanel)

GE currently always show the latest stash, this was expected by mstv in the initial version.
The latest stash is therefore always shown if stashes are shown.
This commit will always include the index and untracked commits.
See https://github.com/gitextensions/gitextensions/pull/10138#issuecomment-1236200583 for a discussion about how that could be changed.

There is no new menu operation to create a stash, use stash button or Ctrl-Alt-Up to create a stash.
Create a stash operate on both Index and WorkTree commits, so there is no natural way to connect this to a revision (but _partial_ commit should be possible to create in RevDiff).
FormStash was improved in #9303 too.

### Performance

Loading revisions in 4.0 is heavily optimized, some performance measurements to ensure there is no regression.
(I have tried to make sure the addition has minimal impact on performance).

My GE repo has 19400 commits and 542 stashes, testing on Win10 with Zen2600, 16GB, SSD.
I consider 542 stashes to be an upper limit to what is reasonable to test with.

If stashes are not displayed, basically one null variable is, adding less than 1 ms to process all revisions. So usecases not showing stashes are not affected.
Enabled but no existing stashes has a few more checks, affects loading with about 9 ms.

Times are approximate, some variation (+/- 30 ms), but relative consistent differences.
Time will be affected by number of cores etc
Refreshing refs is done in three partly parallel steps:

1. Getting refs and checkout
Total step 175 ms (git-for-each-ref 125 ms)

2. Getting Stash commit (10 stash checked for untracked)
Total step 175 ms (git-stash-list 95 ms, git-log <untracked-id> 70 ms)
If this time is higher than step 1 the load time is increasing.
Note that the second Git command for untracked is not needed if reflog is shown.

3. Getting revisions
Total git-log time 600 ms but first revision is delivered and could be processed for display after 140 ms (this is when the spinner is replaced with "Loading" in upper right corner) (as refs are slower, revgrid will be updated 35 ms after this)
(if sorting in the grid is enabled, revisions will be shown with at least a 100 ms delay)
Processing time to find if stashes should be added is about 9 ms.
The GUI update is done in about the time that git-log runs.
The "Loading" label is then removed, but some additional CPU may be needed for selecting a new revision (if filtering changed).

Stashes slow down showing revisions in the grid with about 9 ms plus possible additional time due to parallel processing overhead (but I have not seen any faster ref handling with stash disabled).
This is assuming that the 10 "untracked" takes the same time as getting refs, which in my case is 10, which I also believe is an OK value of untracked to show.
0 untracked limits step 2 with 60 ms (but revisions are not shown earlier), 50 untracked adds 30 ms and all 542 adds 400 ms.

Brief tests with Linux repo in wsl limited to 100000 commits shown with one local branch and 10 stashes.
With so few refs, refs are handled in 150 ms and stash-list in 170 and listing untracked will add 100 ms and 50 ms (processing adding to grid) out of total 4.1 s. So hardly noticeable.

## Screenshots <!-- Remove this section if PR does not change UI -->

With reflogs shown, the change is that labels and menu operations are enabled for all commits.

### Before

Latest stash shown

![image](https://user-images.githubusercontent.com/6248932/183312947-b2455131-227b-497b-bfa2-0d64f81a3e59.png)

### After

If parent commit is not shown, stash is not shown (first is shown though, not updated example)

![image](https://user-images.githubusercontent.com/6248932/183313063-0eab1103-cd6a-4dcd-bd1f-15f1677afa20.png)

Other commits are shown, "untracked" commit only if it is not empty

![image](https://user-images.githubusercontent.com/6248932/183313149-76b55c74-7377-4a37-b4ee-d2b0eae52539.png)

Menu

![image](https://user-images.githubusercontent.com/6248932/189742955-ded2242a-3c92-4fd8-9e08-f157a4d22d97.png)

## Test methodology <!-- How did you ensure quality? -->

A few tests added, some modified.

## Test environment(s) <!-- Remove any that don't apply -->

Git 2.37.1
No other version tested, nothing seen in the Git doc about any compatibility.

## Merge strategy

I agree that the maintainer squash merge this PR (if the commit message is clear).

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).