@mstv 2022-08-24
I have not had a look at the details yet.
Though the current approach does not display (the most recent) stash if it is based on an invisible commit, e.g. "disappears" after a rebase.
The stash index is shown also only if the reflog is shown.