@gerhardol 2022-08-24
Latest stash can always be included (it is explicitly excluded now), but then the latest stash will always look like 
![image](https://user-images.githubusercontent.com/6248932/186434392-8b1d3df4-5e64-4a8e-9f2f-c20d436141d3.png)

But maybe a fixed number of stashes can always be added ("stash@{0} stash@{1} stash@{2} ... stash@{20}") by default.
Then there would be a smaller delay then if stashes have to be queried first (so maybe 20 ms extra instead of 200 ms in my GE clone).

The full info should be seen in the side panel (not implemented).