@gerhardol 2022-08-24
Some measurements:
git-log <stash@{} for 542 stashes> adds about 550 ms to the listing of the first commit. Limiting to 20 adds about 20 ms. (So about 1 ms/commit. So if "force include" is used, the list should be limited. This will require a hard to explain configuration. I canparse about 2000 stashes, then the command line get too long.

To avoid adding the "index" commit and empty untracked, processing time added is about 10 ms, similar handling as for inserting untracked commits