@mstv 2022-08-25
> Latest stash can always be included (it is explicitly excluded now), but then the latest stash will always look like the following, unless there are some funky code to remove the index and empty untracked commits. ![image](https://user-images.githubusercontent.com/6248932/186434392-8b1d3df4-5e64-4a8e-9f2f-c20d436141d3.png) 

That's how it used to be. Some of this info is rarely really needed indeed. Though stashes can be not just one commit. So, alright for me.

> But maybe a fixed number of stashes can always be added ("stash@{0} stash@{1} stash@{2} ... stash@{20}") by default.

Will Git complain if there are fewer stashes?

> Then there would be a smaller delay then if stashes have to be queried first (so maybe 20 ms extra instead of 200 ms in my GE clone).

👍
I think 10 would suffice.