@gerhardol 2022-08-25
> > But maybe a fixed number of stashes can always be added ("stash@{0} stash@{1} stash@{2} ... stash@{20}") by default.
> 
> Will Git complain if there are fewer stashes?

No, that is fine.

> > Then there would be a smaller delay then if stashes have to be queried first (so maybe 20 ms extra instead of 200 ms in my GE clone).
> 
> 👍 I think 10 would suffice.

I have a problem explaining this setting, so not keen on it.
As `rev-parse stash@{0} etc` can take long time, the complete stash listing should probably be done before listing in the grid, always delaying commit presentation.
(Except for reflog when the stashes are added anyway)