@mstv 2022-08-25
> I have a problem explaining this setting, so not keen on it.

OK, so explicitly request just the most recent stash.