@RussKie 2022-08-28
Perhaps worth considering bumping the stash'related menus to the top:
![image](https://user-images.githubusercontent.com/4403806/187076433-e486f20e-e866-4e31-acea-ff1056902968.png)