@gerhardol 2022-09-03
> > I have a problem explaining this setting, so not keen on it.
> 
> OK, so explicitly request just the most recent stash.

Changed to show the latest stash if ShowStashes

If  (currently hidden) and `MinimumStashesAlwaysInGrid` are added in settings, the descriptions could be something like below.
Likely a follow up. Similar with stash in sidepanel.

![image](https://user-images.githubusercontent.com/6248932/188286599-8fff9030-6486-445d-8abb-844075fa4d45.png)

`ShowStashes`
(Checkbox as in _View_ menu)

`MinimumStashesAlwaysInGrid` (default 1)
Stashes in the grid are by default shown in the grid if the parent commit is visible in the revision grid.
(If reflogs are shown, all commits including all stashes are shown.)
This setting will ensure that the stashes are added to the grid.
Note that the _index_ and _untracked_ commits are always included.
Note: Setting this value to a high value may delay the display of commits.

`MaxStashesWithUntrackedFiles` (default 10)
By default only `MinimumStashesAlwaysInGrid` stashes will include the stash commit with untracked files.
This setting will include the untracked commit (unless it is empty) for additional stashes.
Note: Setting this value to a high value may delay the display of commits.