@mstv 2022-09-05
>> After `recovering` the `lost object`, the commit is filtered out somehow as long as RefLog references are shown.

The restored commit "LOST_FOUND_1" (and the index commit) is not returned by
`git -c log.showSignature=false log -z --pretty=format:"%H%T%P%n%at%n%ct%n%aN%n%aE%n%cN%n%cE%n%B"`
neither with nor without `--reflog`.
How is it found and added if **not** showing RefLog references?
The base commit c9b8c5d is only returned *with* `--reflog`. Weird!
 
![image](https://user-images.githubusercontent.com/36601201/188501444-93a59815-b594-497c-b98e-ca2f00a6c02b.png)